Move import screen to a separate WalletImport component. This way it should be easier to apply import functionality on other branches.
